调整初始数值

+++ src/main.cpp
@@ -605,8 +605,8 @@ void setup() {
   if (EEPROM.read(address + 100) != float(233)){ //检测是否是第一次刷入 无意义数值
     EEPROM.put(address + 20, float(11.1)); //初始关电压
     EEPROM.put(address + 30, float(12));  //初始开电压
-    EEPROM.put(address + 40, float(100)); //初始关亮度
-    EEPROM.put(address + 50, float(250)); //初始开亮度
+    EEPROM.put(address + 40, float(300)); //初始关亮度
+    EEPROM.put(address + 50, float(550)); //初始开亮度
     EEPROM.put(address + 100, float(233)); //写入无意义数值 在下次不再初始化
     EEPROM.commit();
   }
@@ -1000,35 +1000,15 @@ String makePage(String title, String contents) {
 
 String urlDecode(String input) {
   String s = input;
-  s.replace("%20", " ");
-  s.replace("+", " ");
-  s.replace("%21", "!");
-  s.replace("%22", "\"");
-  s.replace("%23", "#");
-  s.replace("%24", "$");
-  s.replace("%25", "%");
-  s.replace("%26", "&");
-  s.replace("%27", "\'");
-  s.replace("%28", "(");
-  s.replace("%29", ")");
-  s.replace("%30", "*");
-  s.replace("%31", "+");
-  s.replace("%2C", ",");
-  s.replace("%2E", ".");
-  s.replace("%2F", "/");
-  s.replace("%2C", ",");
-  s.replace("%3A", ":");
-  s.replace("%3A", ";");
-  s.replace("%3C", "<");
-  s.replace("%3D", "=");
-  s.replace("%3E", ">");
-  s.replace("%3F", "?");
-  s.replace("%40", "@");
-  s.replace("%5B", "[");
-  s.replace("%5C", "\\");
-  s.replace("%5D", "]");
-  s.replace("%5E", "^");
-  s.replace("%5F", "-");
-  s.replace("%60", "`");
+  s.replace("%20", " ");s.replace("+", " ");s.replace("%21", "!");
+  s.replace("%22", "\"");s.replace("%23", "#");s.replace("%24", "$");
+  s.replace("%25", "%");s.replace("%26", "&");s.replace("%27", "\'");
+  s.replace("%28", "(");s.replace("%29", ")");s.replace("%30", "*");
+  s.replace("%31", "+");s.replace("%2C", ",");s.replace("%2E", ".");
+  s.replace("%2F", "/");s.replace("%2C", ",");s.replace("%3A", ":");
+  s.replace("%3A", ";");s.replace("%3C", "<");s.replace("%3D", "=");
+  s.replace("%3E", ">");s.replace("%3F", "?");s.replace("%40", "@");
+  s.replace("%5B", "[");s.replace("%5C", "\\");s.replace("%5D", "]");
+  s.replace("%5E", "^");s.replace("%5F", "-");s.replace("%60", "`");
   return s;
 }
